The game is far more functional.  Haven't implemented the tree yet, but thirst is implemented and the map seems to be working correctly.
The reader of the script could spot the map and be able to figure out
how to win, but I'm not sure what I can do to prevent that.

A lot of code to keep someone on the map...  Maybe not quite so much
code is needed for that.

This game is loosely based on The Sands of Egypt for the color computer.